pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "pipecat-ai-flows"
version = "0.0.8"
description = "Conversation Flow management for Pipecat AI applications"
license = { text = "BSD 2-Clause License" }
readme = "README.md"
requires-python = ">=3.10"
keywords = ["pipecat", "conversation", "flows", "state machine", "ai", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Communications :: Conferencing",
    "Topic :: Multimedia :: Sound/Audio",
]
dependencies = [
    "pipecat-ai>=0.0.49",
    "loguru~=0.7.2",
]

[project.urls]
Source = "https://github.com/pipecat-ai/pipecat-flows"
Website = "https://www.pipecat.ai"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100

select = [
    "D", # Docstring rules
]
ignore = ["D212"]
exclude = [
    "examples"
]

[tool.ruff.pydocstyle]
convention = "google"